+++ crates/vm/backends/exec_db.rs
@@ -1,0 +1,36 @@
+use std::collections::HashMap;
+
+use bytes::Bytes;
+use ethrex_common::{
+    types::{AccountInfo, ChainConfig},
+    Address, H256, U256,
+};
+use ethrex_trie::NodeRLP;
+use serde::{Deserialize, Serialize};
+
+/// In-memory EVM database for single execution data.
+///
+/// This is mainly used to store the relevant state data for executing a single block and then
+/// feeding the DB into a zkVM program to prove the execution.
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+pub struct ExecutionDB {
+    /// indexed by account address
+    pub accounts: HashMap<Address, AccountInfo>,
+    /// indexed by code hash
+    pub code: HashMap<H256, Bytes>,
+    /// indexed by account address and storage key
+    pub storage: HashMap<Address, HashMap<H256, U256>>,
+    /// indexed by block number
+    pub block_hashes: HashMap<u64, H256>,
+    /// stored chain config
+    pub chain_config: ChainConfig,
+    /// Encoded nodes to reconstruct a state trie, but only including relevant data ("pruned trie").
+    ///
+    /// Root node is stored separately from the rest as the first tuple member.
+    pub state_proofs: (Option<NodeRLP>, Vec<NodeRLP>),
+    /// Encoded nodes to reconstruct every storage trie, but only including relevant data ("pruned
+    /// trie").
+    ///
+    /// Root node is stored separately from the rest as the first tuple member.
+    pub storage_proofs: HashMap<Address, (Option<NodeRLP>, Vec<NodeRLP>)>,
+}
