+++ crates/vm/backends/mod.rs
@@ -1,3 +1,4 @@
+pub mod exec_db;
 pub mod levm;
 pub mod revm;
